refactor useTopCollections
